requirements.txt
fastapi>=0.113.0,<0.114.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
anthropic>=0.25.0
PyGitHub>=1.59.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
langchain>=0.1.0
langgraph>=0.2.0
langchain-core>=0.3.0 
azure-ai-inference
azure-core
pygments
urllib3>=2.2.1
typing-extensions
# Core Web Framework & Server
fastapi
uvicorn[standard]
websockets
Jinja2
aiofiles

# LLM Integration
google-generativeai

# Configuration & Parsing
python-dotenv
pydantic
json5
langgraph

# Redis (Phase 2.2 - Distributed Queue)
redis>=5.0.0

# File System Monitoring
watchdog>=4.0.0

# Database (Phase 2 - Production Enhancement)
asyncpg>=0.29.0
alembic>=1.13.0
sqlalchemy>=2.0.0